pyproject.toml
[project]
name = "kina"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-mgmt-compute>=34.0.0",
    "azure-mgmt-containerservice>=34.1.0",
    "azure-mgmt-network>=28.1.0",
    "azure-mgmt-resource>=23.3.0",
    "pyyaml>=6.0.2",
    "typer>=0.15.2",
]

[project.scripts]
kina = "kina:app"

[tool.uv]
package = true
